requirements.txt
scikit-learn
numpy
opencv-python
transformers
torch
torchvision
pillow